teams & account image upload solution rectify
