+++ src/main/java/com/ksu/soccerserver/config/SecurityConfig.java
@@ -64,6 +64,7 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
                 .antMatchers(HttpMethod.GET,"/api/accounts/images/**").permitAll() //유저 이미지 허가
                 .antMatchers(HttpMethod.GET,"/api/accounts/**/images/**").permitAll() //유저 이미지 허가
                 .antMatchers(HttpMethod.GET,"/api/teams/images/**").permitAll() //팀 이미지 허가
+                .antMatchers(HttpMethod.GET,"/api/teams/**/images/**").permitAll() //팀 이미지 허가
                 .antMatchers("/api/accounts/**").hasRole("USER") //Account(Modify, Remove, Load) 인증요구
                 .antMatchers("/api/login", "/api/logout").permitAll() //Login/out 허가
                 .antMatchers("/api/find/**").permitAll() //회원정보 찾기 허가
