+++ src/main/java/com/ksu/soccerserver/image/AccountImageController.java
@@ -12,20 +12,15 @@ import javax.servlet.http.HttpServletRequest;
 @RequiredArgsConstructor
 public class AccountImageController {
 
-    private final AccountImageService accountImageService;
+    private final ImageService imageService;
 
     // Image Load
     @GetMapping({"/images/{imageName:.+}", "/{idx}/images/{imageName:.+}"})
     ResponseEntity<?> getImage(@PathVariable(name = "imageName") String imageName,
                                @PathVariable(name = "idx",required = false) Long idx,
                                HttpServletRequest request) {
-        return accountImageService.loadAsResource(imageName, request);
+        return imageService.loadAsResource(imageName, request);
     }
-    // Image 초기화
-//    @PutMapping("/{accountId}/images")
-//    public ResponseEntity<?> removeImage(@PathVariable Long accountId, HttpServletRequest request){
-//        return accountImageService.setuserImage(accountId, request);
-//    }
 
 }
