+++ src/main/java/com/ksu/soccerserver/team/Team.java
@@ -1,7 +1,6 @@
 package com.ksu.soccerserver.team;
 
 import com.fasterxml.jackson.annotation.JsonIdentityInfo;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.ObjectIdGenerators;
 import com.ksu.soccerserver.account.Account;
 import com.ksu.soccerserver.application.ApplicationAccount;
@@ -13,6 +12,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+
 import javax.persistence.*;
 import java.util.HashSet;
 import java.util.Set;
@@ -81,13 +81,11 @@ public class Team {
 
     public void updateTeamInfo(TeamModifyRequest teamModifyRequest){
         this.description = teamModifyRequest.getDescription();
+        this.logopath = teamModifyRequest.getLogopath();
     }
 
     public void joinMember(Account account) {
         accounts.add(account);
     }
-    // Logo 수정
-    public void setLogo(String logo) { this.logopath = logo; }
-
 
 }
