+++ src/main/java/com/ksu/soccerserver/team/dto/TeamRequest.java
@@ -13,6 +13,7 @@ public class TeamRequest {
     private String district;
     private String description;
 
+
     public Team toEntity(Account nowAccount, String image) {
         return Team.builder().name(this.getName())
                 .state(this.getState()).district(this.getDistrict())
